+++ terratest/tests/reverse_dns_function_test.go
@@ -1,14 +1,14 @@
-/*
- * SPDX-FileCopyrightText: The terraform-provider-ipcalc Authors
- * SPDX-License-Identifier: 0BSD
- */
+// Copyright (c) LederWorks
+// SPDX-FileCopyrightText: The terraform-provider-iactools Authors
+// SPDX-License-Identifier: MPL-2.0
 
 package acceptance_test
 
 import (
+	"testing"
+
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/stretchr/testify/assert"
-	"testing"
 )
 
 func TestReverseDnsFunction(t *testing.T) {