+++ CHANGELOG.md
@@ -4,12 +4,14 @@ SPDX-FileCopyrightText: The terraform-provider-iactools Authors
 SPDX-License-Identifier: MPL-2.0
 -->
 
-## 0.2.0 (Unreleased)
+## 0.2.0 (Released)
+
 FEATURES:
 - Added CodeQL workflow
 - Added Dependabot automerge workflow
 - Added REUSE workflow
 - Added Verify workflow
+- Added GitVersion workflow
 - Added reverse_dns function
 
 ENHANCEMENTS:
@@ -18,6 +20,10 @@ ENHANCEMENTS:
 - Updated README.md
 - Updated CODE_OF_CONDUCT.md
 - Improved documentation
+- Improved testing
+
+BUG FIXES:
+- inverse_cidr now calculates CIDR ranges with a Trie Graph
 
 ## 0.1.0 (Unreleased)